BUG make sure to delete files properly and test for errors

+++ eastlake/steps/delete_meds.py
@@ -26,15 +26,18 @@ class DeleteMeds(Step):
             if tilename in self.config["save_tilenames"]:
                 continue
 
-            # Get meds filenames
-            meds_files = stash.get_filepaths("meds_files", tilename)
-            for m in meds_files:
-                self.logger.error("removing meds file %s" % m)
-                os.remove(m)
+            meds_files = stash.get_filepaths("meds_files", tilename, keyerror=False)
+            if meds_files is not None:
+                for m in meds_files:
+                    if os.path.isfile(m):
+                        self.logger.error("removing meds file %s" % m)
+                        os.remove(m)
 
-            meds_files = stash.get_filepaths("pizza_cutter_meds_files", tilename)
-            for m in meds_files:
-                self.logger.error("removing pizza-cutter meds file %s" % m)
-                os.remove(m)
+            meds_files = stash.get_filepaths("pizza_cutter_meds_files", tilename, keyerror=False)
+            if meds_files is not None:
+                for m in meds_files:
+                    if os.path.isfile(m):
+                        self.logger.error("removing pizza-cutter meds file %s" % m)
+                        os.remove(m)
 
         return 0, stash
